[translate]: Added delete endpoint tests. Improved delete endpoint and defined return codes better. Added some delete endpoint security tests.

+++ appcomposer/user/__init__.py
@@ -1,6 +1,6 @@
 import datetime
 from appcomposer import db
-from appcomposer.models import User
+from appcomposer.models import User, App, AppVar
 
 
 def create_user(login, name, password):
@@ -18,8 +18,19 @@ def remove_user(user):
     """
     Deletes a user from the DB.
 
-    @param user: User object. It is *NOT* a login string.
+    @param user: User object. It is *NOT* a login string. If None then nothing is done.
     """
+    if user is None:
+        return None
+
+    # Get every AppVar and App for the user.
+    apps = App.query.filter_by(owner=user).all()
+    for app in apps:
+        appvars = AppVar.query.filter_by(app=app).all()
+        for var in appvars:
+            db.session.delete(var)
+        db.session.delete(app)
+
     db.session.delete(user)
     db.session.commit()
     return None
